Refactor buckets into a module.
