+++ .env.sample
@@ -1,0 +1,2 @@
+GITHUB_API_TOKEN=
+
