+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2022-10-31 20:37:54 UTC using RuboCop version 1.36.0.
+# on 2022-11-08 23:16:17 UTC using RuboCop version 1.36.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -170,7 +170,7 @@ Style/ClassAndModuleChildren:
   Exclude:
     - 'spec/support/without_data_files.rb'
 
-# Offense count: 3
+# Offense count: 4
 # This cop supports unsafe autocorrection (--autocorrect-all).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: always, always_true, never
