+++ lib/github/pull_requests.rb
@@ -2,50 +2,14 @@
 
 module GitHub
   class PullRequests < Items
+    include GitHub::Buckets
+
     def initialize(arr_or_options)
       super arr_or_options, GitHub::PullRequest
     end
 
     def contributors
       @contributors ||= GitHub::Contributors.new(map(&:user))
-    end
-
-    def members
-      buckets[:members] || []
-    end
-
-    def contractors
-      buckets[:contractors] || []
-    end
-
-    def external
-      buckets[:external] || []
-    end
-
-    def unknown
-      buckets[:unknown] || []
-    end
-
-    def [](bucket)
-      buckets[bucket] || []
-    end
-
-    def all
-      to_a
-    end
-
-    def internal
-      members.concat(contractors)
-    end
-
-    def all_external
-      external
-    end
-
-    def all_external_percent
-      return 0 unless all.any?
-
-      ((all_external.size.to_f / all.size) * 100).to_i
     end
 
     def buckets
