+++ lib/tools.rb
@@ -7,6 +7,7 @@ require 'ruby-progressbar'
 require 'chronic'
 require 'dotiw'
 require 'redcarpet'
+require_relative 'github/buckets'
 require_relative 'github/rate_limited'
 require_relative 'github/item'
 require_relative 'github/progress'
